% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocen.R
\name{geocen}
\alias{geocen}
\title{Geocoding function for Census API}
\usage{
geocen(
  oneline = NULL,
  street = NULL,
  city = NULL,
  state = NULL,
  zip = NULL,
  separate = FALSE,
  vintage = "4",
  layers = "8",
  benchmark = "Public_AR_Current",
  format = "json"
)
}
\arguments{
\item{oneline}{One or single line address combining street, city, state, and zip for geocoding.}

\item{street}{Street text for address provided. Defaults to NULL for "oneline" preference.}

\item{city}{City text for address provided. Defaults to NULL for "oneline" preference.}

\item{state}{State text for address provided. Defaults to NULL for "oneline" preference.}

\item{zip}{Zip text for address provided. Defaults to NULL for "oneline" preference.}

\item{separate}{If TRUE, geocoder automatically uses separate address field arguments in request. Defaults to FALSE.}

\item{vintage}{Defaults to "4".}

\item{layers}{Returns BASENAME info from layers listed in the Census WMS. Default is only "8" for addition Census Tract info.}

\item{benchmark}{Defaults to "Public_AR_Current".}

\item{format}{Defaults to "json".}
}
\description{
This function allows you geocode an address and return extra layers using the Census Bureau's geocoding API
}
\examples{
geocen()
}
\keyword{census}
\keyword{geocoder}
